written add folder tests
