+++ tests/folder_tests/page_component.py
@@ -1,0 +1,19 @@
+import urllib.parse as urlparse
+
+
+class Page(object):
+    BASE_URL = 'https://e.mail.ru/settings/folders'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+
+    def open(self):
+        url = urlparse.urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
+
+
+class Component(object):
+    def __init__(self, driver):
+        self.driver = driver
